finishing changes
